fix: projections behavior when removing a removed entity

+++ src/RoadRegistry.Wms.Projections/RoadSegmentRecordProjection.cs
@@ -9,6 +9,7 @@ using BackOffice.Messages;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.Connector;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.SqlStreamStore;
 using Microsoft.EntityFrameworkCore;
+using NetTopologySuite.Triangulate;
 using Schema;
 using Syndication.Schema;
 
@@ -380,6 +381,10 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WmsContext>
     private static async Task RemoveRoadSegment(RoadSegmentRemoved roadSegmentRemoved, WmsContext context)
     {
         var roadSegmentRecord = await context.RoadSegments.FindAsync(roadSegmentRemoved.Id).ConfigureAwait(false);
+        if (roadSegmentRecord == null)
+        {
+            throw new InvalidOperationException($"RoadSegmentRecord with id {roadSegmentRemoved.Id} is not found");
+        }
 
         context.RoadSegments.Remove(roadSegmentRecord);
     }
